Remove uppercase characters from project name
It was sometimes written as "mingw-w64", sometimes as "MinGW-w64"
(I'd guess for historic reasons coming from MinGW)
In many cases where lowercase is required such as package names it's
always "mingw-w64". Let's make things simpler by just going with
lowercase where possible, i.e. no special uppercase characters.
(I've asked on IRC and they were fine with it)

I've left it capitalized if it is at the start of a sentence in
most cases for now, as I wasn't sure what's the right thing to do there.

While at it also update the logo tag line to include arm64.
And set a logo width/height to avoid a layout shift on loading.